% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloX}
\alias{helloX}
\title{Prints "Hello" + str param in funny colors}
\usage{
helloX(str)
}
\arguments{
\item{str}{text variable}
}
\description{
Prints "Hello" + str param in funny colors
}
