% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalUniStgtt.R
\docType{data}
\name{pal_uni_sgtt}
\alias{pal_uni_sgtt}
\title{Primary colors of the University of Stuttgart}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://www.beschaeftigte.uni-stuttgart.de/uni-services/oeffentlichkeitsarbeit/corporate-design/cd-dateien/Uni_Stuttgart_CD-Manual.pdf}{CD manual}.
}
\usage{
pal_uni_sgtt
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
Primary color 1 "WEISS"
RGB 255 / 255 / 255
HEX #FFFFFF
}
\details{
Primary color 2 "ANTHRAZIT"
PANTONE 432 U
CMYK 50 / 35 / 25 / 70
RGB 62 / 68 / 76
HEX #323232

NOTE:
RGB != HEX, we work with HEX!
#323232 => RGB 50 / 50 / 50
RGB 62 / 68 / 76 => #3E444C

Primary color 3 "MITTELBLAU"
PANTONE 293 U
CMYK 100 / 70 / 0 / 0
RGB 0 / 81 / 158
HEX #004191

NOTE:
RGB != HEX, we work with HEX!
#004191 => RGB 0 / 65 / 145
RGB 0 / 81 / 158 => #00519E


Secondary color "HEllBlaU"
PANTONE 2985 U
CMYK 70 / 0 / 0 / 0
RGB 0 / 190 / 255
HEX #00BEFF

\code{uni_sgtt} provides the three primary colors and one secondary color
of the \href{https://www.uni-stuttgart.de/en}{University of Stuttgart}, Germany.
}
\examples{
pal_uni_sgtt
unikn::seecol(pal_uni_sgtt, main = "Primary colors of the University of Stuttgart")  # view color palette hex colors

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-05-31.
}
\concept{contributed color palettes}
\keyword{datasets}
