% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalUniStgtt.R
\docType{data}
\name{pal_uni_sgtt_hex}
\alias{pal_uni_sgtt_hex}
\title{Primary colors of the University of Stuttgart}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://www.beschaeftigte.uni-stuttgart.de/uni-services/oeffentlichkeitsarbeit/corporate-design/cd-dateien/Uni_Stuttgart_CD-Manual.pdf}{CD manual}.
}
\usage{
pal_uni_sgtt_hex
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_sgtt} provides the three primary colors and one secondary color
of the \href{https://www.uni-stuttgart.de/en}{University of Stuttgart}, Germany.
}
\examples{
pal_uni_sgtt_hex
unikn::seecol(pal_uni_sgtt_hex, main = "Primary colors of the University of Stuttgart HEX")  # view color palette hex colors

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other contributed color palettes:
\code{\link{pal_uni_sgtt_rgb}}
}
\author{
\strong{unicol}, 2023-05-31.
}
\concept{contributed color palettes}
\keyword{datasets}
