% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalUniStgtt.R
\docType{data}
\name{uni_sgtt}
\alias{uni_sgtt}
\title{Primary colors of the University of Stuttgart}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://www.beschaeftigte.uni-stuttgart.de/uni-services/oeffentlichkeitsarbeit/corporate-design/cd-dateien/Uni_Stuttgart_CD-Manual.pdf}{CD manual}.
}
\usage{
uni_sgtt
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_sgtt} provides the three primary colors and one secondary color
of the \href{https://www.uni-stuttgart.de/en}{University of Stuttgart}, Germany.
}
\details{
Primary color "WEISS"
RGB 255 / 255 / 255
WEB #FFFFFF

Primary color "ANTHRAZIT"
PANTONE 432 U

CMYK 50 / 35 / 25 / 70

RGB 62 / 68 / 76

WEB #323232

Primary color "MITTELBLAU"
PANTONE 293 U
CMYK 100 / 70 / 0 / 0
RGB 0 / 81 / 158
WEB #004191

Secondary color "HEllBlaU"
PANTONE 2985 U
CMYK 70 / 0 / 0 / 0
RGB 0 / 190 / 255
WEB #00BEFF
}
\examples{
uni_sgtt
seecol(uni_sgtt, main = "Primary colors of the University of Stuttgart")  # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-05-31.
}
\concept{contributed color palettes}
\keyword{datasets}
